+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.10.4]
+
+### Added
+
+- Tests to cover introspection
+
+### Fixed
+
+- Issue related to introspection token
+
 ## [0.10.3]
 
 ### Added